fix type casts

+++ api/machine_info_test.go
@@ -21,7 +21,7 @@ func TestMachineInfo(t *testing.T) {
 		Name:   "TestMachineInfo",
 		Image:  "ubuntu/trusty",
 		Driver: "lxd",
-		UserID: testToken.Claims.(jwt.MapClaims)["id"].(int64),
+		UserID: int64(testToken.Claims.(jwt.MapClaims)["id"].(float64)),
 	}
 
 	if err := testDB.Insert(&machine); err != nil {
